Temporarily disable python3 job
We want to include the python3 zuul jobs in the tests repo.
For that we need to make sure the job is not defined anywhere
else. We are removing all the references from
ansible-config_template, openstack-ansible-plugins in all
branches, before re-introducing them as jobs in the tests
repo, and using them in each repo project.yaml

Change-Id: Ic5899a1aeee2183858766547e623fc605e52d54f