Fix py3 unicode issue
unicode is nolonger a type in py3 - instead the str type is used for
unicode, and the bytes type would cover the str type from py2.

To avoid having to version the code, we should try to use unicode,
except the NameError and continue as though it is py3 if that fails.

Additionally, this patch adds a test that will fail if you revert the
config_template.py back to it's original in py3.

Closes-bug: 1763422
Change-Id: Ifda972caada27ade2d80f77b3df70568406226ff