Allow config_template to be loaded as a role
Currently when this repo loaded as a role instead of copied to
plugins paths, config_template tasks fail because:
FAILED! => {"msg": "module (config_template) is missing interpreter line"}

This is because ansible searches for action plugins in the action_plugins/
path of roles and plays instead of action/. See
https://docs.ansible.com/ansible/2.5/plugins/action.html#enabling-action-plugins
for information on this.

By simply symlinking action_plugins -> action we can make this work
in either scenario.

Change-Id: I59625e358e220b83d4a2580ac5f99bc11751ebdd