Fix ordering for --diff mode setup
Due to the ordering of the vars passed to the _get_diff_data mode,
--diff mode would only work with --check activated, otherwise the file
used didn't exist and you would get an error similar to:

https://github.com/ceph/ceph-ansible/issues/2742

This PR allows this to work by configuring the diff data before running
the copy module, so the file exists.

Additionally, we should run with --diff mode enabled by default on tests
for this repository to ensure any functional issues with diff mode are
caught.

Change-Id: I020a181731c3ee87f04933f8c62213ea12cfa3a4