Add fallback option when stating the source file
This change will now fall back to using the ansible core module
`stat` to stat a file when the local stat function fails for a
permissions error. This is being done because there is an issue
with action plugins not respecting "become: true" on local files
being written over a local connection.

Change-Id: I646f0688f1f4bf3cddaef9114dac51c2a807b25b
Signed-off-by: Kevin Carter <kecarter@redhat.com>