Update documentation about installation methods
config_template historically was published on PyPi during times before
collections were a thing. This was one of the installations methods
that is not relevant or preffered since collections are here.

That patch aims to update documentation about ways of plugin
installation as well as adding deprecation notice of PyPi.

In the futher releases we will drop publishing on PyPi.

Change-Id: Ie1cd7b305b4d82fd0fbf4568ef20ff3d0236c049